fix/ complete breakdown of large files
